Use a big integer field for invoice costs

+++ app/grandchallenge/invoices/models.py
@@ -182,7 +182,7 @@ class Invoice(models.Model, FieldChangeMixin):
     compute_costs_euros = models.PositiveIntegerField(
         help_text="The capacity reservation in Euros"
     )
-    compute_cost_euro_millicents = models.PositiveIntegerField(
+    compute_cost_euro_millicents = models.PositiveBigIntegerField(
         help_text="The utilized compute costs in Euro millicents (cached from utilizations)",
         default=0,
     )
